0.47-9
Fix DMA Alignment

+++ Core/Src/led.c
@@ -12,7 +12,7 @@
 extern TIM_HandleTypeDef htim3;
 extern bool	isLEDsendpulse;
 uint8_t	LEDColor[LED_COUNT];	// coded LED color value
-uint8_t	LEDPulse[TOTAL_BITS];	// Data formed PWM width send to LED
+uint16_t	LEDPulse[TOTAL_BITS];	// Data formed PWM width send to LED
 uint8_t	LEDTimer[LED_COUNT];	// Individual LED Timer Counter
 
 /**
